added presentationDetents and Radius to UIKitBackend

+++ Sources/SwiftCrossUI/Values/PresentationDetent.swift
@@ -7,10 +7,12 @@ public enum PresentationDetent: Sendable, Hashable {
     case large
 
     /// A detent at a custom fractional height of the available space.
+    /// falling back to medium on iOS 15
     /// - Parameter fraction: A value between 0 and 1 representing the fraction of available height.
     case fraction(Double)
 
     /// A detent at a specific fixed height in pixels.
-    /// - Parameter height: The height in pixels.
-    case height(Int)
+    /// falling back to medium on iOS 15
+    /// - Parameter height: The height
+    case height(Double)
 }
